moves EntityTagChecker outside package
