+++ src/main/java/de/rabea/request/Directory.java
@@ -79,5 +79,4 @@ public class Directory {
         });
         return files;
     }
-
 }
